+++ ttnn/cpp/ttnn/operations/data_movement/tilize_with_val_padding/device/tilize_with_val_padding_single_core_program_factory.cpp
@@ -1,0 +1,239 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#include "tilize_with_val_padding_single_core_program_factory.hpp"
+
+#include <math.h>
+
+#include "ttnn/operations/cb_utils.hpp"
+#include "ttnn/operations/math.hpp"
+#include "ttnn/operation.hpp"
+#include "ttnn/operations/core/work_split/work_split_tilize.hpp"
+#include <tt-metalium/constants.hpp>
+#include <tt-metalium/util.hpp>
+#include <tt-metalium/host_api.hpp>
+#include <tt-metalium/allocator.hpp>
+#include "ttnn/operations/data_movement/tilize_with_val_padding/tilize_with_val_padding_common.hpp"
+#include <tt-metalium/work_split.hpp>
+
+using namespace tt::constants;
+using namespace tt::tt_metal;
+
+namespace ttnn::operations::data_movement::detail {
+
+uint32_t get_packed_value(const Tensor tensor, const ttnn::PadValue pad_value) {
+    return std::visit(
+        [&tensor](auto&& pad_value) {
+            using T = std::decay_t<decltype(pad_value)>;
+            if constexpr (std::is_same_v<T, float>) {
+                if (tensor.dtype() == DataType::BFLOAT16) {
+                    bfloat16 bfloat_pad_value = bfloat16((pad_value));
+                    return pack_two_bfloat16_into_uint32({bfloat_pad_value, bfloat_pad_value});
+                } else {
+                    TT_FATAL(
+                        tensor.dtype() == DataType::FLOAT32 or tensor.dtype() == DataType::UINT32,
+                        "only supporting bfloat16, float32, and uint32");
+                    return (uint32_t)((pad_value));
+                }
+            } else if constexpr (std::is_same_v<T, uint32_t>) {
+                if (tensor.dtype() == DataType::BFLOAT16) {
+                    bfloat16 bfloat_pad_value = bfloat16((float)(pad_value));
+                    return pack_two_bfloat16_into_uint32({bfloat_pad_value, bfloat_pad_value});
+                } else {
+                    TT_FATAL(
+                        tensor.dtype() == DataType::FLOAT32 or tensor.dtype() == DataType::INT32 or
+                            tensor.dtype() == DataType::UINT32,
+                        "only supporting bfloat16, float32, and int32/uint32");
+                    return ((pad_value));
+                }
+            } else {
+                TT_THROW("type not supported");
+            }
+        },
+        pad_value);
+}
+
+tt::tt_metal::operation::ProgramWithCallbacks tilize_with_val_padding_single_core(
+    const Tensor& a, Tensor& output, const ttnn::PadValue pad_value) {
+    tt::tt_metal::Program program = tt::tt_metal::CreateProgram();
+
+    CoreRange core({0, 0}, {0, 0});
+
+    // This should allocate a DRAM buffer on the device
+    tt::tt_metal::IDevice* device = a.device();
+
+    tt::tt_metal::Buffer* src0_buffer = a.buffer();
+
+    tt::DataFormat input_cb_data_format = tt::tt_metal::datatype_to_dataformat_converter(a.dtype());
+    uint32_t input_single_tile_size = tt::tt_metal::detail::TileSize(input_cb_data_format);
+
+    tt::DataFormat output_cb_data_format = tt::tt_metal::datatype_to_dataformat_converter(output.dtype());
+    uint32_t output_single_tile_size = tt::tt_metal::detail::TileSize(output_cb_data_format);
+
+    int32_t num_tiles = output.physical_volume() / TILE_HW;
+
+    auto true_input_shape = a.padded_shape();
+    auto true_output_shape = output.padded_shape();
+
+    auto input_w = true_input_shape.rank() >= 4 ? true_input_shape[-4] : 1;
+    auto input_z = true_input_shape.rank() >= 3 ? true_input_shape[-3] : 1;
+    auto input_y = true_input_shape.rank() >= 2 ? true_input_shape[-2] : 1;
+    auto input_x = true_input_shape[-1];
+
+    auto output_w = true_output_shape.rank() >= 4 ? true_output_shape[-4] : 1;
+    auto output_z = true_output_shape.rank() >= 3 ? true_output_shape[-3] : 1;
+    auto output_y = true_output_shape.rank() >= 2 ? true_output_shape[-2] : 1;
+    auto output_x = true_output_shape[-1];
+
+    uint32_t unpadded_row_size_bytes = input_x * a.element_size();  // Assuming bfloat16 dataformat
+    uint32_t padded_row_size_bytes = output_x * a.element_size();   // Assuming bfloat16 dataformat
+
+    constexpr uint32_t alignment = 32;
+
+    uint32_t num_tiles_in_row = output_x / TILE_WIDTH;
+    // Ensure we don't intrude into storage space
+    uint32_t max_l1_size =
+        a.device()->l1_size_per_core() / 2 - a.device()->allocator()->get_base_allocator_addr(HalMemType::L1);
+    // Memory usage is 2 CBs of width W, plus buffer of size alignment + (W * datum size)
+    uint32_t max_X = (max_l1_size - alignment) / (a.element_size() * TILE_HEIGHT * 2 + a.element_size());
+    uint32_t max_tiles = max_X / TILE_WIDTH;
+
+    // Currently need the number of tiles in a row to be divisible by tiles in a block
+    uint32_t num_tiles_per_block = 1;
+    if (num_tiles_in_row <= max_tiles) {
+        num_tiles_per_block = num_tiles_in_row;
+    } else {
+        for (uint32_t n_t = max_tiles; n_t > 0; n_t--) {
+            if (num_tiles_in_row % n_t == 0) {
+                num_tiles_per_block = n_t;
+                break;
+            }
+        }
+    }
+
+    uint32_t block_width = num_tiles_per_block * TILE_WIDTH;
+    uint32_t block_row_size = block_width * a.element_size();
+    uint32_t num_blocks_w_output = padded_row_size_bytes / block_row_size;
+    uint32_t num_blocks_w_input = unpadded_row_size_bytes / block_row_size;
+
+    // Leftover size if input is not divisible by block size
+    uint32_t block_row_leftover_size = unpadded_row_size_bytes - num_blocks_w_input * block_row_size;
+
+    // Number of blocks that differ between input and output
+    const uint32_t num_blocks_w_diff = num_blocks_w_output - num_blocks_w_input - (block_row_leftover_size > 0 ? 1 : 0);
+
+    const uint32_t padded_Y_diff_blocks = (output_y - input_y) / TILE_HEIGHT * num_blocks_w_output;
+    const uint32_t padded_Z_diff_blocks = (output_z - input_z) * output_y / TILE_HEIGHT * num_blocks_w_output;
+    const uint32_t padded_W_diff_blocks =
+        (output_w - input_w) * output_z * output_y / TILE_HEIGHT * num_blocks_w_output;
+    const uint32_t num_leftover_Y = input_y - input_y / TILE_HEIGHT * TILE_HEIGHT;
+
+    tt::tt_metal::Buffer* dst_buffer = output.buffer();
+    TT_ASSERT(dst_buffer != nullptr, "Output buffer should be allocated on device!");
+
+    uint32_t src0_cb_index = 0;
+    uint32_t num_input_tiles = num_tiles_per_block;
+    assert(num_input_tiles > 0);
+    tt::tt_metal::CircularBufferConfig src0_cb_config =
+        tt::tt_metal::CircularBufferConfig(
+            num_input_tiles * input_single_tile_size, {{src0_cb_index, input_cb_data_format}})
+            .set_page_size(src0_cb_index, input_single_tile_size);
+    auto cb_src0 = tt::tt_metal::CreateCircularBuffer(program, core, src0_cb_config);
+
+    uint32_t output_cb_index = tt::CBIndex::c_16;
+    uint32_t num_output_tiles = num_tiles_per_block;
+    tt::tt_metal::CircularBufferConfig cb_output_config =
+        tt::tt_metal::CircularBufferConfig(
+            num_output_tiles * output_single_tile_size, {{output_cb_index, output_cb_data_format}})
+            .set_page_size(output_cb_index, output_single_tile_size);
+    auto cb_output = tt::tt_metal::CreateCircularBuffer(program, core, cb_output_config);
+
+    uint32_t packed_pad_value = get_packed_value(a, pad_value);
+    uint32_t tile_row_size_bytes = (a.dtype() == DataType::BFLOAT16) ? 64 : 128;
+
+    const std::array reader_kernel_args = {
+        src0_buffer->address(),
+        input_w,
+        padded_W_diff_blocks,
+        input_z,
+        padded_Z_diff_blocks,
+        input_y,
+        padded_Y_diff_blocks,
+        num_leftover_Y,
+        input_x,
+        unpadded_row_size_bytes,
+        padded_row_size_bytes,
+        packed_pad_value,
+        num_blocks_w_input,
+        num_blocks_w_output,
+        num_blocks_w_diff,
+        block_row_size,
+        block_row_leftover_size};
+
+    // Reader compile-time args
+    uint32_t src0_is_dram = src0_buffer->buffer_type() == tt::tt_metal::BufferType::DRAM ? 1 : 0;
+    uint32_t stick_size = unpadded_row_size_bytes;
+    uint32_t stick_size_is_power_of_two = is_power_of_two_at_least_32(stick_size);
+    uint32_t log2_stick_size = stick_size_is_power_of_two ? (uint32_t)log2(stick_size) : 0;
+    std::vector<uint32_t> reader_compile_time_args = {
+        src0_is_dram, stick_size_is_power_of_two, log2_stick_size, tile_row_size_bytes};
+
+    // Tilized reader
+    tt::tt_metal::KernelHandle unary_reader_kernel_id = tt::tt_metal::CreateKernel(
+        program,
+        "ttnn/cpp/ttnn/operations/data_movement/tilize_with_val_padding/device/kernels/dataflow/"
+        "reader_unary_pad_dims_split_rows.cpp",
+        core,
+        tt::tt_metal::ReaderDataMovementConfig(reader_compile_time_args));
+
+    // Tilized writer
+    uint32_t out_is_dram = dst_buffer->buffer_type() == tt::tt_metal::BufferType::DRAM ? 1 : 0;
+    tt::tt_metal::KernelHandle unary_writer_kernel_id = tt::tt_metal::CreateKernel(
+        program,
+        "ttnn/cpp/ttnn/operations/eltwise/unary/device/kernels/dataflow/writer_unary_interleaved_start_id.cpp",
+        core,
+        tt::tt_metal::WriterDataMovementConfig({output_cb_index, out_is_dram}));
+
+    std::vector<uint32_t> compute_kernel_args = {
+        uint32_t(num_tiles / num_tiles_per_block), uint32_t(num_tiles_per_block)};
+
+    auto tilize_kernel_id = tt::tt_metal::CreateKernel(
+        program,
+        "ttnn/cpp/ttnn/deprecated/tt_dnn/kernels/compute/tilize.cpp",
+        core,
+        tt::tt_metal::ComputeConfig{.compile_args = compute_kernel_args});
+
+    tt::tt_metal::SetRuntimeArgs(program, unary_reader_kernel_id, core, reader_kernel_args);
+
+    tt::tt_metal::SetRuntimeArgs(
+        program, unary_writer_kernel_id, core, {dst_buffer->address(), (uint32_t)num_tiles, 0});
+
+    auto override_runtime_args_callback = [reader_kernel_id = unary_reader_kernel_id,
+                                           writer_kernel_id = unary_writer_kernel_id](
+                                              const void* operation,
+                                              Program& program,
+                                              const std::vector<Tensor>& input_tensors,
+                                              const std::vector<std::optional<const Tensor>>& optional_tensors,
+                                              const std::vector<Tensor>& output_tensors) {
+        auto src_buffer = input_tensors.at(0).buffer();
+
+        auto dst_buffer = output_tensors.at(0).buffer();
+
+        CoreCoord core = {0, 0};
+
+        {
+            auto& runtime_args = GetRuntimeArgs(program, reader_kernel_id, core);
+            runtime_args[0] = src_buffer->address();
+        }
+
+        {
+            auto& runtime_args = GetRuntimeArgs(program, writer_kernel_id, core);
+            runtime_args[0] = dst_buffer->address();
+        }
+    };
+
+    return {std::move(program), override_runtime_args_callback};
+}
+
+}  // namespace ttnn::operations::data_movement::detail
